Add Example / Test code
This is the code that was used to verify the functionality of the Floral
Bonnet.  You'll need a few external repositories to fully test
everything (and submodules seems like a mess to do for something this
simple).  Just run gitclone.sh and snag the respective repositories.

To see how things are run take a look at commands.txt