Add starlight commands & BME280 & TSL2561 driver
This pulls in the drivers for the BME280 and the TSL2561 directly
instead of relying on the upstream location, particularly since they
haven't been updated since 2018, and the license is compatible here.

This also adds in basic commands to get everything working on the
BeagleV Starlight board which has a compatible 2x20 pin header.

Currently the luma examples are disabled as trying to get the SSD1306
working is proving obnoxious because, apparently, the "only" board
with accessible GPIOs seems to be the rPi, which is great, except
there's lots of boards with spidev and gpios SOOOOOOO, this is
currently untested, and slightly obnoxious

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>