Rotate oled screen 90deg counter-clockwise
This is basically a manufacturing optimization to move the under side
components of the oled screen away from the pi header.  There's a cap
that comes a bit close to the header and it's troublesome.  This shifts
the screen around and allows for just some kapton tape on the oled
screen to alleviate any possible touching / arcing.

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>